requirements.txt
asgiref==3.7.2
Django==4.2.2
django-simple-history==3.3.0
djangorestframework==3.14.0
Pillow==9.5.0
pydantic==1.10.9
pypng==0.20220715.0
pytz==2023.3
qrcode==7.4.2
segno==1.5.2
sqlparse==0.4.4
typing_extensions==4.6.3
